Fix git graph arguments
The previous implementation of the `git graph` script had a bug where
the `--date-order` argument was not being correctly passed to the `git
log` command. This commit fixes this issue by ensuring that the
`--date-order` argument is correctly passed to the `git log` command
when it is specified.

Additionally, this commit refactors the script to use a more consistent
style for passing arguments to the `git log` command.